Safari-609.1.11.1/LayoutTests/http/tests/security/mixedContent/import-insecure-script-in-iframe-expected.txt

CONSOLE MESSAGE: line 1: [blocked] The page at https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-import.html was not allowed to run insecure content from http://127.0.0.1:8080/security/mixedContent/resources/script.js.

This test loads a secure iframe that attempt to import() a module script via HTTP. Since the iframe's content is of a secure origin, importing a module script via HTTP should result in a mixed content error.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.data is "TypeError: Importing a module script failed."
PASS successfullyParsed is true

TEST COMPLETE
